// hr-frontend/src/lib/api/generated/model/createContractRequest.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { CreateContractRequestSalaryBasis } from './createContractRequestSalaryBasis';

export interface CreateContractRequest {
  contractTypeId: number;
  contractTemplateId: number;
  /**
     * @minLength 0
     * @maxLength 50
     */
  contractNumber: string;
  contractDate: string;
  startDate: string;
  endDate?: string;
  /** @minimum 0.01 */
  salary?: number;
  salaryBasis: CreateContractRequestSalaryBasis;
  /**
     * @minLength 3
     * @maxLength 10
     */
  salaryCurrency: string;
  /** @minimum 0.01 */
  hourlyRate?: number;
  /** @minimum 0.01 */
  maxMonthlyHours?: number;
  /** @minimum 0.01 */
  workingHoursPerWeek?: number;
  /** @minimum 0.01 */
  workingHoursPerMonth?: number;
  /**
     * @minimum 0
     * @maximum 3650
     */
  probationPeriodDays?: number;
  fulltime: boolean;
  /**
     * @minLength 0
     * @maxLength 255
     */
  projectName?: string;
  /**
     * @minLength 0
     * @maxLength 255
     */
  externalEmployerName?: string;
  externalLeaveStartDate?: string;
  externalLeaveEndDate?: string;
  /**
     * @minimum 0
     * @maximum 3650
     */
  noticePeriodDays?: number;
  /**
     * @minLength 0
     * @maxLength 4000
     */
  notes?: string;
}